+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/Common.kt
@@ -83,6 +83,38 @@ sealed interface StickerType {
     }
 }
 
+@Serializable(StickerFormat.Serializer::class)
+sealed interface StickerFormat {
+    val type: String
+
+    @Serializable
+    object Static : StickerFormat { override val type: String = "static" }
+    @Serializable
+    object Animated : StickerFormat { override val type: String = "animated" }
+    @Serializable
+    object Video : StickerFormat { override val type: String = "video" }
+    @Serializable
+    data class Unknown(override val type: String = "custom_emoji") : StickerFormat
+
+    object Serializer : KSerializer<StickerFormat> {
+        override val descriptor: SerialDescriptor = String.serializer().descriptor
+
+        override fun deserialize(decoder: Decoder): StickerFormat {
+            return when (val type = decoder.decodeString()) {
+                Static.type -> Static
+                Animated.type -> Animated
+                Video.type -> Video
+                else -> Unknown(type)
+            }
+        }
+
+        override fun serialize(encoder: Encoder, value: StickerFormat) {
+            encoder.encodeString(value.type)
+        }
+
+    }
+}
+
 val usernameRegex = Regex("@[\\w\\d_]+")
 
 val degreesLimit = 1 .. 360
@@ -388,6 +420,8 @@ const val webmStickerField = "webm_sticker"
 const val oldChatMemberField = "old_chat_member"
 const val newChatMemberField = "new_chat_member"
 const val stickerTypeField = "sticker_type"
+const val stickerFormatField = "sticker_format"
+const val needsRepaintingField = "needs_repainting"
 
 const val okField = "ok"
 const val captionField = "caption"